requirements.txt
flask
opencv-python
numpy
torch
faiss-cpu
cn-clip
pillow
transformers
tqdm
nudenet
